pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "foamclient"
version = "0.1.4"
description = ""
authors = [
    {name = "Jun Zhu", email="zhujun981661@gmail.com"},
]
license = {file = "LICENSE"}
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    'msgpack',
    'numpy',
    'pyzmq',
    'hiredis',
    'redis',
    'avro>=1.11.2'
]
requires-python = ">=3.9"

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
]

[project.urls]
repository = "https://github.com/zhujun98/foamclient"
